backend/requirements.win-postgres.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
pydantic>=2
pydantic-settings>=2
python-dotenv>=1

email-validator==2.*

python-multipart==0.*

passlib[bcrypt]==1.7.*

python-jose[cryptography]==3.*

PyJWT==2.*

cryptography==43.*

python-multipart==0.*

passlib[bcrypt]==1.7.*

requests==2.*

google-api-python-client==2.*

google-auth==2.*

google-auth-httplib2==0.*

google-auth-oauthlib==1.*

google-api-python-client==2.*

google-auth==2.*

google-auth-httplib2==0.*

google-auth-oauthlib==1.*

requests==2.*

google-api-python-client==2.*

google-auth==2.*

google-auth-httplib2==0.*

google-auth-oauthlib==1.*
